backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic[email]==2.5.0
pyjwt==2.8.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==41.0.7
qrcode[pil]==7.4.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
